+++ packages/create-app/templates/solidjs-static/vite.config.js
@@ -1,25 +1,36 @@
 import { defineConfig, loadEnv } from 'vite';
 import solidPlugin from 'vite-plugin-solid';
 import { createHtmlPlugin } from 'vite-plugin-html';
-import { createBlockletPlugin } from 'vite-plugin-blocklet';
+import createWssHmrPlugin from 'vite-plugin-wss-hmr';
 
-export default ({ mode }) => {
+export default defineConfig(async ({ mode }) => {
   const envMap = loadEnv(mode, process.cwd(), '');
   const port = process.env.BLOCKLET_PORT || 3000;
 
-  return defineConfig({
+  const whenDev = mode === 'development';
+
+  let mountPoint = process.env.BLOCKLET_DEV_MOUNT_POINT || '';
+
+  if (mountPoint && !mountPoint.endsWith('/')) {
+    mountPoint = `${mountPoint}/`;
+  }
+
+  const base = whenDev ? mountPoint : process.env.BASE_URL || '/';
+
+  return {
+    base,
     plugins: [
       solidPlugin(),
       createHtmlPlugin({
         minify: true,
         inject: {
           data: {
-            base: process.env.BASE_URL || '/',
+            base,
             title: envMap.APP_TITLE,
           },
         },
       }),
-      createBlockletPlugin(),
+      await createWssHmrPlugin(),
     ],
     build: {
       target: 'esnext',
@@ -28,5 +39,5 @@ export default ({ mode }) => {
     server: {
       port,
     },
-  });
-};
+  };
+});