+++ packages/create-app/templates/vue-static/vite.config.js
@@ -1,29 +1,40 @@
 import { defineConfig, loadEnv } from 'vite';
 import vue from '@vitejs/plugin-vue';
 import { createHtmlPlugin } from 'vite-plugin-html';
-import { createBlockletPlugin } from 'vite-plugin-blocklet';
+import createWssHmrPlugin from 'vite-plugin-wss-hmr';
 
 // https://vitejs.dev/config/
-export default ({ mode }) => {
+export default defineConfig(async ({ mode }) => {
   const envMap = loadEnv(mode, process.cwd(), '');
   const port = process.env.BLOCKLET_PORT || 3000;
 
-  return defineConfig({
+  const whenDev = mode === 'development';
+
+  let mountPoint = process.env.BLOCKLET_DEV_MOUNT_POINT || '';
+
+  if (mountPoint && !mountPoint.endsWith('/')) {
+    mountPoint = `${mountPoint}/`;
+  }
+
+  const base = whenDev ? mountPoint : process.env.BASE_URL || '/';
+
+  return {
+    base,
     plugins: [
       vue(),
       createHtmlPlugin({
         minify: true,
         inject: {
           data: {
-            base: process.env.BASE_URL || '/',
+            base,
             title: envMap.APP_TITLE,
           },
         },
       }),
-      createBlockletPlugin(),
+      await createWssHmrPlugin(),
     ],
     server: {
       port,
     },
-  });
-};
+  };
+});