+++ scripts/bump-version.mjs
@@ -1,11 +1,19 @@
 /* eslint-disable no-console */
 import { execSync } from 'child_process';
 import { $, chalk, fs } from 'zx';
+import batchModifyDepsVersion from './batch-modify-deps-version.mjs';
 
 // or use pnpm to bump version: `pnpm -r --filter {packages/*, themes/*} -- pnpm version`
-execSync('bumpp package.json packages/*/package.json', { stdio: 'inherit' });
+execSync('bumpp package.json packages/*/package.json plugins/*/package.json', { stdio: 'inherit' });
 
 const { version } = await fs.readJSON('package.json');
+
+// modify the /create-app/templates plugin version
+await batchModifyDepsVersion({
+  dirPath: './packages/create-app/templates',
+  depList: await fs.readdirSync(`./plugins`), // scan the plugins directory to get the dep list
+  version: `^${version}`,
+});
 
 console.log(`\nNow you can make adjustments to ${chalk.cyan('CHANGELOG.md')}. Then press enter to continue.`);