+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.11 (七月 08, 2022)
+
+- feat: bump-version script can update the create-app/templates plugin version
+- feat: support the templates of vue/solidjs/svelte(use vite) dev in the component mode
+
 ## 0.4.10 (七月 07, 2022)
 
 - feat: support the templates of vue(use @vue/cli) dev in the component mode